# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/work-packages/{workPackageId}": {
      "get": {
        "operationId": "WorkPackagesController_get",
        "parameters": [
          {
            "name": "workPackageId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Read one WorkPackage. Cross-tenant and cross-project ids are denied without leakage.",
        "tags": [
          "operations"
        ]
      },
      "patch": {
        "operationId": "WorkPackagesController_update",
        "parameters": [
          {
            "name": "workPackageId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update WorkPackage fields. Status, ownership, and Deliverable association use dedicated endpoints. CAS expectedVersion is required.",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/work-packages/{workPackageId}/context": {
      "get": {
        "operationId": "WorkPackageTraceabilityController_context",
        "parameters": [
          {
            "name": "workPackageId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Authorized WorkPackage context sections. Gate links are read-only.",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/milestones": {
      "get": {
        "operationId": "MilestonesController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List Milestones. status is derived; recordedStatus is the stored explicit state.",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "MilestonesController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a Milestone in PLANNED. AT_RISK / MISSED are derived, not client-set.",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{projectId}/milestones/{milestoneId}": {
      "get": {
        "operationId": "MilestonesController_get",
        "parameters": [
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "planning"
        ]
      },
      "patch": {
        "operationId": "MilestonesController_update",
        "parameters": [
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update Milestone fields. Status is not patchable.",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{projectId}/milestones/{milestoneId}/achieve": {
      "post": {
        "operationId": "MilestonesController_achieve",
        "parameters": [
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Explicitly achieve a Milestone. Completing linked Tasks does not achieve it.",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{projectId}/milestones/{milestoneId}/cancel": {
      "post": {
        "operationId": "MilestonesController_cancel",
        "parameters": [
          {
            "name": "milestoneId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Cancel a planned Milestone",
        "tags": [
          "planning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Amber API",
    "description": "Amber Modular Monolith foundation API. Formal Exception is the sole Gate bypass. Closed 0.2A catalog only.",
    "version": "1.6.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "amber_session"
      }
    },
...
  }
}